correcciones de consulta tp3

+++ tp3/ejercicio4/AnalizadorArbol.java
@@ -27,7 +27,7 @@ public class AnalizadorArbol {
                     }
                 }
                 else {
-                    maxPromedio = Math.max(maxPromedio, acumulador/cant);
+                    maxPromedio = Math.max(maxPromedio, (double)acumulador/(double)cant);
                     
                     if (!cola.isEmpty()){
                         cola.enqueue(null);
